Rename SportBike file to PascalCase
